NerPipeline + TokenClassificationPipeline
