+++ src/transformers/pipelines/__init__.py
@@ -29,7 +29,6 @@ from transformers.pipelines.base import (
     ConversationalPipeline,
     CsvPipelineDataFormat,
     JsonPipelineDataFormat,
-    NerPipeline,
     PipedPipelineDataFormat,
     Pipeline,
     PipelineDataFormat,
@@ -40,8 +39,6 @@ from transformers.pipelines.base import (
     TableQuestionAnsweringArgumentHandler,
     TableQuestionAnsweringPipeline,
     Text2TextGenerationPipeline,
-    TokenClassificationArgumentHandler,
-    TokenClassificationPipeline,
     TranslationPipeline,
     get_default_model,
     get_framework,
@@ -50,6 +47,11 @@ from transformers.pipelines.feature_extraction import FeatureExtractionPipeline
 from transformers.pipelines.fill_mask import FillMaskPipeline
 from transformers.pipelines.text_classification import TextClassificationPipeline
 from transformers.pipelines.text_generation import TextGenerationPipeline
+from transformers.pipelines.token_classification import (
+    NerPipeline,
+    TokenClassificationArgumentHandler,
+    TokenClassificationPipeline,
+)
 from transformers.pipelines.zero_shot_classification import (
     ZeroShotClassificationArgumentHandler,
     ZeroShotClassificationPipeline,
